Updated project structure | Fixed up components
