+++ src/app/components/AboutUI/index.js
@@ -1,7 +1,4 @@
-import { Links } from "../../../globals/consts";
-import { toPublicPath } from "../../../utils/route-helper";
-import CustomTitle from "../CustomTitle";
-import styles from "./styles";
+import template from "./template";
 
 export default class AboutUI extends HTMLElement {
   static tagName = "about-ui";
@@ -9,27 +6,7 @@ export default class AboutUI extends HTMLElement {
   constructor() {
     super();
     this.attachShadow({ mode: "open" });
-    this._render();
-  }
-
-  _render() {
-    this.shadowRoot.innerHTML = `
-      ${styles}
-      <div class="about-container">
-        <div class="image">
-          <img src="${toPublicPath("/images/landing/about.png")}" loading="lazy" alt="About Picture" />
-        </div>
-        <div class="about">
-          <${CustomTitle.tagName}
-            variant="${CustomTitle.VARIANT.H3}"
-            text="Rully Ihza Mahendra"
-            size="${CustomTitle.SIZE.MEDIUM}">
-          </${CustomTitle.tagName}>
-
-          <p>I am a <a href="${Links.googleCertified}">Google Certified Associate Android Developer</a>, based in Semarang, Indonesia. And also one of 3000 selected Mobile Programming students at <a href="${Links.bangkit}">Bangkit Academy 2021 led by Google, Tokopedia, Gojek & Traveloka</a> graduated with distinction status. I really love programming, especially in Android Dev. And I mainly and mostly work with Kotlin and Flutter. However, I'm crazy with web tech as well, just for my hobbies.</p>
-        </div>
-      </div>
-    `;
+    this.shadowRoot.innerHTML = template;
   }
 }
