+++ src/app/components/ContainerApp/template.js
@@ -1,0 +1,8 @@
+import styles from "./styles";
+
+const template = /* html */`
+  ${styles}
+  <div class="container-app"></div>
+`;
+
+export default template;
