+++ src/app/components/CustomAlert/index.js
@@ -1,168 +1,12 @@
-import logger from "../../../utils/logger";
 import { isOnMobileScreen } from "../../../utils/ui/viewport-helpers";
-import Dom from "../../core/Dom";
-import styles from "./styles";
-import CustomButton from "../CustomButton";
+import ButtonText from "../ButtonText";
+import { Callbacks, CustomAlertBuilder } from "./builder";
+import template from "./template";
 
-class CustomAlertDetail {
-  constructor(title, message, type, size) {
-    this.title = title;
-    this.message = message;
-    this.type = type;
-    this.size = size;
-  }
-}
-
-const Callbacks = {
-  confirm: null,
-  cancel: null,
-};
-
-class CustomAlertBuilder {
-  static #detail = null;
-
-  static #title;
-
-  static #message;
-
-  static #type;
-
-  static #size;
-
-  static #cancelable = false;
-
-  static #confirmText = "Oke";
-
-  static #cancelText = "Cancel";
-
-  static isShowing = false;
-
-  static getDetail() {
-    if (CustomAlertBuilder.#detail === null) {
-      CustomAlertBuilder.#detail = new CustomAlertDetail("", "", "success");
-    }
-    return CustomAlertBuilder.#detail;
-  }
-
-  static setTitle(title = "") {
-    CustomAlertBuilder.#title = title;
-    return CustomAlertBuilder;
-  }
-
-  static setMessage(message = "") {
-    CustomAlertBuilder.#message = message;
-    return CustomAlertBuilder;
-  }
-
-  static setType(type = "succes") {
-    CustomAlertBuilder.#type = type;
-    return CustomAlertBuilder;
-  }
-
-  static setSize(size = "medium") {
-    CustomAlertBuilder.#size = size;
-    return CustomAlertBuilder;
-  }
-
-  static setConfirm(text, callback = null) {
-    CustomAlertBuilder.#confirmText = text;
-    if (callback === null) {
-      Callbacks.confirm = () => {};
-    } else {
-      Callbacks.confirm = callback;
-    }
-    return CustomAlertBuilder;
-  }
-
-  static setCancel(text, callback = null) {
-    CustomAlertBuilder.#cancelText = text;
-    Callbacks.cancel = callback;
-    return CustomAlertBuilder;
-  }
-
-  static setCancelable(cancelable = false) {
-    CustomAlertBuilder.#cancelable = cancelable;
-    return CustomAlertBuilder;
-  }
-
-  static build() {
-    CustomAlertBuilder.#detail = new CustomAlertDetail(
-      CustomAlertBuilder.#title,
-      CustomAlertBuilder.#message,
-      CustomAlertBuilder.#type,
-      CustomAlertBuilder.#size,
-    );
-    return CustomAlertBuilder;
-  }
-
-  static show() {
-    if (CustomAlertBuilder.#detail === null) {
-      logger.error("CustomAlert not created yet!");
-      return;
-    }
-
-    if (CustomAlertBuilder.isShowing) {
-      logger.error("CustomAlert is already showing!");
-      return;
-    }
-
-    // show
-    const alertContainer = Dom.createElement({
-      tagName: "div",
-      styles: {
-        position: "fixed",
-        top: "0",
-        right: "0",
-        left: "0",
-        bottom: "0",
-        "z-index": "var(--z-alert)",
-        backgroundColor: "rgba(82, 82, 82, .2)",
-        display: "flex",
-        justifyContent: "center",
-      },
-    });
-
-    alertContainer.innerHTML = `
-      <custom-alert
-        title="${CustomAlertBuilder.#detail.title}"
-        message="${CustomAlertBuilder.#detail.message}"
-        type="${CustomAlertBuilder.#detail.type}"
-        size="${CustomAlertBuilder.#detail.size}"
-        cancel-text="${CustomAlertBuilder.#cancelText}"
-        confirm-text="${CustomAlertBuilder.#confirmText}"
-        ${Callbacks.confirm === null ? "single-confirm" : ""}
-        >
-      </custom-alert>
-    `;
-
-    if (document.body.style.overflowY !== "hidden") {
-      document.body.style.overflowY = "hidden";
-    }
-
-    document.body.appendChild(alertContainer);
-    CustomAlertBuilder.isShowing = true;
-
-    // remove handler
-    alertContainer.addEventListener("click", (e) => {
-      if (!CustomAlertBuilder.#cancelable) return;
-
-      const clickedNodeName = e.target.parentNode.nodeName;
-
-      if (clickedNodeName === "BODY") { // only container
-        if (document.body.style.overflowY === "hidden") {
-          document.body.style.overflowY = "auto";
-        }
-        if (CustomAlertBuilder.isShowing) alertContainer.remove();
-        CustomAlertBuilder.isShowing = false;
-      }
-    });
-
-    this.#detail = null;
-  }
-}
-
-class CustomAlert extends HTMLElement {
+export default class CustomAlert extends HTMLElement {
   static tagName = "custom-alert";
+
+  static Builder = CustomAlertBuilder;
 
   static TYPE = {
     SUCCESS: "success",
@@ -228,7 +72,7 @@ class CustomAlert extends HTMLElement {
   }
 
   connectedCallback() {
-    this.shadowRoot.querySelector(`${CustomButton.tagName}:first-child`)
+    this.shadowRoot.querySelector(`${ButtonText.tagName}:first-child`)
       .addEventListener("click", () => {
         // run callback
         if (Callbacks.cancel && typeof Callbacks.cancel === "function") {
@@ -238,8 +82,8 @@ class CustomAlert extends HTMLElement {
         this.removeAlert();
       });
 
-    if (this.shadowRoot.querySelector(`${CustomButton.tagName}:nth-child(2)`)) {
-      this.shadowRoot.querySelector(`${CustomButton.tagName}:nth-child(2)`)
+    if (this.shadowRoot.querySelector(`${ButtonText.tagName}:nth-child(2)`)) {
+      this.shadowRoot.querySelector(`${ButtonText.tagName}:nth-child(2)`)
         .addEventListener("click", () => {
           // run callback
           if (Callbacks.confirm && typeof Callbacks.confirm === "function") {
@@ -270,17 +114,17 @@ class CustomAlert extends HTMLElement {
     const mediumSize = "45%";
     const smallSize = "35%";
 
-    let buttonsSize = CustomButton.SIZE.MEDIUM;
+    let buttonsSize = ButtonText.SIZE.MEDIUM;
 
     if (this.#size === CustomAlert.SIZE.BIG) {
       this.style.width = isMobile ? xlargeSize : largeSize;
-      buttonsSize = CustomButton.SIZE.MEDIUM;
+      buttonsSize = ButtonText.SIZE.MEDIUM;
     } else if (this.#size === CustomAlert.SIZE.MEDIUM) {
       this.style.width = isMobile ? largeSize : mediumSize;
-      buttonsSize = isMobile ? CustomButton.SIZE.SMALL : CustomButton.SIZE.MEDIUM;
+      buttonsSize = isMobile ? ButtonText.SIZE.SMALL : ButtonText.SIZE.MEDIUM;
     } else {
       this.style.width = isMobile ? largeSize : smallSize;
-      buttonsSize = CustomButton.SIZE.SMALL;
+      buttonsSize = ButtonText.SIZE.SMALL;
     }
 
     this.style.marginTop = "20vh";
@@ -304,53 +148,42 @@ class CustomAlert extends HTMLElement {
 
     if (this.#singleConfirm) {
       buttons = `
-        <${CustomButton.tagName}
+        <${ButtonText.tagName}
         text="${this.#cancelText}"
         size="${buttonsSize}"
         bg-color="${this.#singleConfirm && !this.#singleBgColor ? this.#singleBgColor : "grey"}"
         color="var(--white-color)"
         title="Confirm"
-        ></${CustomButton.tagName}>
+        ></${ButtonText.tagName}>
       `;
     } else {
       buttons = `
-      <${CustomButton.tagName}
+      <${ButtonText.tagName}
         text="${this.#cancelText}"
         size="${buttonsSize}"
         bg-color="${this.#singleConfirm && !this.#singleBgColor ? this.#singleBgColor : "grey"}"
         color="var(--white-color)"
         title="Confirm"
-        ></${CustomButton.tagName}>
-      <${CustomButton.tagName}
+        ></${ButtonText.tagName}>
+      <${ButtonText.tagName}
         text="${this.#confirmText}"
         size="${buttonsSize}"
         bg-color="var(--accent-color)"
         color="var(--white-color)"
         title="Confirm"
-        ></${CustomButton.tagName}>
+        ></${ButtonText.tagName}>
       `;
     }
 
-    this.shadowRoot.innerHTML = `
-      ${styles}
-      <div class="alert-container">
-        <p class="icon">${icon}</p>
-        <p class="title">${this.#title}</p>
-        <p class="message">${this.#message}</p>
-
-        <div class=button-container>
-          ${buttons}
-        </div>
-      </div>
-    `;
+    this.shadowRoot.innerHTML = template(
+      this.#title,
+      this.#message,
+      icon,
+      buttons,
+    );
   }
 }
 
 if (!customElements.get(CustomAlert.tagName)) {
   customElements.define(CustomAlert.tagName, CustomAlert);
 }
-
-export {
-  CustomAlert,
-  CustomAlertBuilder,
-};
