--- src/app/components/CustomContainer/styles.js
@@ -1,3 +1,0 @@
-import { css } from "../../core/Style";
-
-export default css``;
