+++ src/app/components/FooterApp/index.js
@@ -1,5 +1,4 @@
-import styles from "./styles";
-import { Social, Strings } from "../../../globals/consts";
+import template from "./template";
 
 export default class FooterApp extends HTMLElement {
   static tagName = "footer-app";
@@ -7,36 +6,7 @@ export default class FooterApp extends HTMLElement {
   constructor() {
     super();
     this.attachShadow({ mode: "open" });
-    this._render();
-  }
-
-  _render() {
-    const yearNow = new Date().getFullYear();
-
-    this.shadowRoot.innerHTML = `
-      ${styles}
-      <footer>
-        <p class="appreciation">${Strings.ThankYou}</p>
-        <p class="connect-with-me">${Strings.GetConnectionWithMe}</p>
-        <div class="social">
-          <a href="${Social.facebook}" target="_blank" title="facebook">
-            <box-icon type='logo' name='facebook' color="var(--white-color)"></box-icon>
-          </a>
-          <a href="${Social.instagram}" target="_blank" title="instagram">
-            <box-icon type='logo' name='instagram' color="var(--white-color)"></box-icon>
-          </a>
-          <a href="${Social.twitter}" target="_blank" title="twitter">
-            <box-icon type='logo' name='twitter' color="var(--white-color)"></box-icon>
-          </a>
-        </div>
-        <p class="built-by-me">
-          ${Strings.BuiltCopyright("<box-icon type='solid' name='heart' color='var(--white-color)' size='16px'></box-icon>")}
-        </p>
-        <p class="copyright">
-          ${Strings.Copyright(yearNow)}
-        </p>
-      </footer>
-    `;
+    this.shadowRoot.innerHTML = template;
   }
 }
