+++ src/app/components/Loading/template.js
@@ -1,0 +1,12 @@
+import styles from "./styles";
+
+const template = /* html */`
+  ${styles}
+  <div class="loading-wrapper">
+    <svg class="spinner" viewBox="0 0 50 50">
+      <circle class="path" cx="25" cy="25" r="20" fill="none" stroke-width="5"></circle>
+    </svg>
+  </div>
+`;
+
+export default template;
