+++ src/app/components/ProjectItem/template.js
@@ -1,0 +1,13 @@
+import styles from "./styles";
+
+const template = (headingVariant, title, imagePath, url) => /* html */`
+  ${styles}
+  <a class="project-container" href="${url}">
+    <div class="loading-preview">
+      <img src="${imagePath}" alt="${title}" loading="lazy" />
+    </div>
+    <${headingVariant} class="title">${title}</${headingVariant}>
+  </a>
+`;
+
+export default template;
