+++ src/app/components/SkillsUI/index.js
@@ -1,7 +1,6 @@
-import { toPublicPath } from "../../../utils/route-helper";
-import styles from "./styles";
-import CustomTitle from "../CustomTitle";
+import TitleText from "../TitleText";
 import { isOnMobileScreen } from "../../../utils/ui/viewport-helpers";
+import template from "./template";
 
 export default class SkillsUI extends HTMLElement {
   static tagName = "skills-ui";
@@ -31,41 +30,14 @@ export default class SkillsUI extends HTMLElement {
 
   constructor() {
     super();
-    this.attachShadow({ mode: "open" });
-    this._render();
-  }
+    const shadow = this.attachShadow({ mode: "closed" });
 
-  _render() {
-    const align = isOnMobileScreen() ? CustomTitle.ALIGN.CENTER : CustomTitle.ALIGN.START;
+    const align = isOnMobileScreen() ? TitleText.ALIGN.CENTER : TitleText.ALIGN.START;
 
-    this.shadowRoot.innerHTML = `
-      ${styles}
-
-      <${CustomTitle.tagName}
-        variant="${CustomTitle.VARIANT.H3}"
-        align="${align}"
-        text="Professional Skills"
-        size="${CustomTitle.SIZE.MEDIUM}">
-      </${CustomTitle.tagName}>
-
-      <div class="container-skills">
-        <div class="skills-detail">
-          <p class="skills-description">I am working with some of programming skills and all of my projects I've ever done built on top of them. Such as:</p>
-          ${this.#skillsData.map((skill) => (`
-            <div class="skills-data">
-              <div class="skills-data-description">
-                ${skill.icon}
-                <span class="skills-data-name">${skill.name}</span>
-              </div>
-              <div class="skills-data-bar"></div>
-            </div>
-          `)).join("")}
-        </div>
-        <div class="skills-illustration">
-          <img src="${toPublicPath("/images/landing/skills.jpg")}" />
-        </div>
-      </div>
-    `;
+    shadow.innerHTML = template(
+      this.#skillsData,
+      align,
+    );
   }
 }
