+++ src/app/components/SkillsUI/styles.js
@@ -1,6 +1,6 @@
 import { css } from "../../core/Style";
 
-export default css`
+const styles = /* css */`
   img {
     width: 100%;
     border-radius: 8px;
@@ -71,3 +71,5 @@ export default css`
     }
   }
 `;
+
+export default css(styles);
