+++ src/app/components/TitleText/styles.js
